+++ mrjob/conf.py
@@ -1,5 +1,6 @@
 # Copyright 2009-2012 Yelp
 # Copyright 2013 David Marin
+# Copyright 2015 Yelp
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,7 +13,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
 """"mrjob.conf" is the name of both this module, and the global config file
 for :py:mod:`mrjob`.
 """
@@ -20,7 +20,7 @@ import glob
 import json
 import logging
 import os
-from itertools import chain
+import os.path
 
 # yaml is nice to have, but we can fall back on JSON if need be
 try:
@@ -110,8 +110,7 @@ class OptionStore(dict):
             raise KeyError(key)
 
 
-### READING AND WRITING mrjob.conf ###
-
+### finding config files ###
 
 def find_mrjob_conf():
     """Look for :file:`mrjob.conf`, and return its path. Places we look:
@@ -143,7 +142,18 @@ def find_mrjob_conf():
         return None
 
 
+# TODO: rename and hide this function in v0.5.0
+
 def real_mrjob_conf_path(conf_path=None):
+    """Return the path of a single conf file. If *conf_path* is ``False``,
+    return ``None``, and if it's ``None``, return :py:func:`find_mrjob_conf`.
+    Otherwise, expand environment variables and ``~`` in *conf_path* and
+    return it.
+
+    Confusingly, this function doesn't actually return a "real" path according
+    to ``os.path.realpath()``; it just resolves environment variables and
+    ``~``.
+    """
     if conf_path is False:
         return None
     elif conf_path is None:
@@ -152,13 +162,139 @@ def real_mrjob_conf_path(conf_path=None):
         return expand_path(conf_path)
 
 
+### !clear tag ###
+
+class ClearedValue(object):
+    """Wrap a value tagged with !clear in mrjob.conf"""
+
+    def __init__(self, value):
+        self.value = value
+
+    def __eq__(self, other):
+        if isinstance(other, ClearedValue):
+            return self.value == other.value
+        else:
+            return False
+
+    def __hash__(self):
+        return hash(self.value)
+
+    def __repr__(self):
+        return '%s(%s)' % (self.__class__.__name__, repr(self.value))
+
+
+def _cleared_value_constructor(loader, node):
+    # tried construct_object(), got an unconstructable recursive node warning
+    if isinstance(node, yaml.MappingNode):
+        value = loader.construct_mapping(node)
+    elif isinstance(node, yaml.ScalarNode):
+        # resolve null as None, not u'null'
+        value = yaml.safe_load(node.value)
+    elif isinstance(node, yaml.SequenceNode):
+        value = loader.construct_sequence(node)
+    else:
+        raise TypeError
+
+    return ClearedValue(value)
+
+
+def _load_yaml_with_clear_tag(stream):
+    """Like yaml.safe_load(), but everything with a !clear tag before it
+    will be wrapped in ClearedValue()."""
+    loader = yaml.SafeLoader(stream)
+    loader.add_constructor('!clear', _cleared_value_constructor)
+    try:
+        return loader.get_single_data()
+    finally:
+        if hasattr(loader, 'dispose'):  # it doesn't in PyYAML 3.09
+            loader.dispose()
+
+
+def _cleared_value_representer(dumper, data):
+    if not isinstance(data, ClearedValue):
+        raise TypeError
+    node = dumper.represent_data(data.value)
+    node.tag = '!clear'
+    return node
+
+
+def _dump_yaml_with_clear_tags(data, stream=None, **kwds):
+    class ClearedValueSafeDumper(yaml.SafeDumper):
+        pass
+
+    ClearedValueSafeDumper.add_representer(
+        ClearedValue, _cleared_value_representer)
+
+    return yaml.dump_all([data], stream, Dumper=ClearedValueSafeDumper, **kwds)
+
+
+def _fix_clear_tags(x):
+    """Recursively resolve :py:class:`ClearedValue` wrappers so that
+    ``ClearedValue(...)`` can only wrap values in dicts (and in the top-level
+    value we return).
+
+    In dicts, we treat ``ClearedValue(k): v`` or
+    ``ClearedValue(k): ClearedValue(v)`` as equivalent to
+    ``k: ClearedValue(v)``. ``ClearedValue(k): v1`` overrides ``k: v2``.
+
+    In lists, any ClearedValue wrappers are simply stripped.
+    """
+    _fix = _fix_clear_tags
+
+    if isinstance(x, list):
+        return [_fix(_strip_clear_tag(item)) for item in x]
+
+    elif isinstance(x, dict):
+        d = dict((_fix(k), _fix(v)) for k, v in x.items())
+
+        # handle cleared keys
+        for k, v in list(d.items()):
+            if isinstance(k, ClearedValue):
+                del d[k]
+                d[_strip_clear_tag(k)] = ClearedValue(_strip_clear_tag(v))
+
+        return d
+
+    elif isinstance(x, ClearedValue):
+        return ClearedValue(_fix(x.value))
+
+    else:
+        return x
+
+
+def _resolve_clear_tags_in_list(items):
+    """Create a list from *items*. If we encounter a :py:class:`ClearedValue`,
+    unwrap it and ignore previous values. Used by ``combine_*()`` functions
+    to combine lists of values.
+    """
+    result = []
+
+    for item in items:
+        if isinstance(item, ClearedValue):
+            result = [item.value]
+        else:
+            result.append(item)
+
+    return result
+
+
+def _strip_clear_tag(v):
+    """remove the clear tag from the given value."""
+    if isinstance(v, ClearedValue):
+        return v.value
+    else:
+        return v
+
+
+### reading mrjob.conf ###
+
 def conf_object_at_path(conf_path):
     if conf_path is None:
         return None
 
     with open(conf_path) as f:
         if yaml:
-            return yaml.safe_load(f)
+            return _fix_clear_tags(_load_yaml_with_clear_tag(f))
         else:
             try:
                 return json.load(f)
@@ -176,8 +312,8 @@ def conf_object_at_path(conf_path):
 def load_opts_from_mrjob_conf(runner_alias, conf_path=None,
                               already_loaded=None):
     """Load a list of dictionaries representing the options in a given
-    mrjob.conf for a specific runner. Returns ``[(path, values)]``. If
-    conf_path is not found, return [(None, {})].
+    mrjob.conf for a specific runner, resolving includes. Returns
+    ``[(path, values)]``. If *conf_path* is not found, return ``[(None, {})]``.
 
     :type runner_alias: str
     :param runner_alias: String identifier of the runner type, e.g. ``emr``,
@@ -185,8 +321,18 @@ def load_opts_from_mrjob_conf(runner_alias, conf_path=None,
     :type conf_path: str
     :param conf_path: location of the file to load
     :type already_loaded: list
-    :param already_loaded: list of :file:`mrjob.conf` paths that have already
-                           been loaded
+    :param already_loaded: list of real (according to ``os.path.realpath()``)
+                           conf paths that have already
+                           been loaded (used by
+                           :py:func:`load_opts_from_mrjob_confs`).
+
+    .. versionchanged:: 0.4.6
+        Relative ``include:`` paths are relative to the real (after resolving
+        symlinks) path of the including conf file
+
+    .. versionchanged:: 0.4.6
+        This will only load each config file once, even if it's referenced
+        from multiple paths due to symlinks.
     """
     conf_path = real_mrjob_conf_path(conf_path)
     conf = conf_object_at_path(conf_path)
@@ -197,8 +343,15 @@ def load_opts_from_mrjob_conf(runner_alias, conf_path=None,
     if already_loaded is None:
         already_loaded = []
 
-    already_loaded.append(conf_path)
+    # don't load same conf file twice
+    real_conf_path = os.path.realpath(conf_path)
 
+    if real_conf_path in already_loaded:
+        return []
+    else:
+        already_loaded.append(real_conf_path)
+
+    # get configs for our runner out of conf file
     try:
         values = conf['runners'][runner_alias] or {}
     except (KeyError, TypeError, ValueError):
@@ -210,39 +363,55 @@ def load_opts_from_mrjob_conf(runner_alias, conf_path=None,
         if isinstance(includes, string_types):
             includes = [includes]
 
-        for include in includes:
-            if include in already_loaded:
-                log.warning('%s tries to recursively include %s! (Already'
-                         ' included:  %s)' % (
-                             conf_path, conf['include'],
-                             ', '.join(already_loaded)))
-            else:
-                inherited.extend(
-                    load_opts_from_mrjob_conf(
-                        runner_alias, include, already_loaded))
+        # handle includes in reverse order so that include order takes
+        # precedence over inheritance
+        for include in reversed(includes):
+            # make include relative to (real) conf_path (see #1166)
+            include = os.path.join(os.path.dirname(real_conf_path), include)
+
+            inherited = load_opts_from_mrjob_conf(
+                runner_alias, include, already_loaded) + inherited
+
     return inherited + [(conf_path, values)]
 
 
 def load_opts_from_mrjob_confs(runner_alias, conf_paths=None):
     """Load a list of dictionaries representing the options in a given
     list of mrjob config files for a specific runner. Returns
-    ``[(path, values)]``. If a path is not found, use (None, {}) as its value.
+    ``[(path, values), ...]``. If a path is not found, use ``(None, {})`` as
+    its value.
+
     If *conf_paths* is ``None``, look for a config file in the default
-    locations.
+    locations (see :py:func:`find_mrjob_conf`).
 
     :type runner_alias: str
     :param runner_alias: String identifier of the runner type, e.g. ``emr``,
                          ``local``, etc.
     :type conf_paths: list or ``None``
     :param conf_path: locations of the files to load
+
+    .. versionchanged:: 0.4.6
+        This will only load each config file once, even if it's referenced
+        from multiple paths due to symlinks.
     """
     if conf_paths is None:
         return load_opts_from_mrjob_conf(runner_alias, find_mrjob_conf())
     else:
-        return chain(*[
-            load_opts_from_mrjob_conf(runner_alias, path)
-            for path in conf_paths])
+        # don't include conf files that were loaded earlier in conf_paths
+        already_loaded = []
 
+        # load configs in reversed order so that order of conf paths takes
+        # precedence over inheritance
+        results = []
+
+        for path in reversed(conf_paths):
+            results = load_opts_from_mrjob_conf(
+                runner_alias, path, already_loaded=already_loaded) + results
+
+        return results
+
+
+### writing mrjob.conf ###
 
 def dump_mrjob_conf(conf, f):
     """Write out configuration options to a file.
@@ -260,7 +429,7 @@ def dump_mrjob_conf(conf, f):
     :param f: a file object to write to (e.g. ``open('mrjob.conf', 'w')``)
     """
     if yaml:
-        yaml.safe_dump(conf, f, default_flow_style=False)
+        _dump_yaml_with_clear_tags(conf, f, default_flow_style=False)
     else:
         json.dump(conf, f, indent=2)
     f.flush()
@@ -270,6 +439,11 @@ def dump_mrjob_conf(conf, f):
 
 # combiners generally consider earlier values to be defaults, and later
 # options to override or add on to them.
+
+# combiners assume that the list of values passed to them has already been
+# passed through _fix_clear_tags() (that is, the only place ClearedValue
+# appears is values in dicts).
+
 
 def combine_values(*values):
     """Return the last value in *values* that is not ``None``.
@@ -288,12 +462,24 @@ def combine_lists(*seqs):
 
     Generally this is used for a list of commands we want to run; the
     "default" commands get run before any commands specific to your job.
+
+    .. versionchanged:: 0.4.6
+       Strings, bytes, and non-sequence objects (e.g. numbers) are treated as
+       single-item lists.
     """
     result = []
 
     for seq in seqs:
-        if seq:
-            result.extend(seq)
+        if seq is None:
+            continue
+
+        if isinstance(seq, (bytes, string_types)):
+            result.append(seq)
+        else:
+            try:
+                result.extend(seq)
+            except:
+                result.append(seq)
 
     return result
 
@@ -317,16 +503,6 @@ def combine_cmds(*cmds):
         return list(cmd)
 
 
-def combine_cmd_lists(*seqs_of_cmds):
-    """Concatenate the given commands into a list. Ignore ``None`` values,
-    and parse strings with :py:func:`shlex.split`.
-
-    Returns a list of lists (each sublist contains the command plus arguments).
-    """
-    seq_of_cmds = combine_lists(*seqs_of_cmds)
-    return [combine_cmds(cmd) for cmd in seq_of_cmds]
-
-
 def combine_dicts(*dicts):
     """Combine zero or more dictionaries. Values from dicts later in the list
     take precedence over values earlier in the list.
@@ -337,19 +513,34 @@ def combine_dicts(*dicts):
 
     for d in dicts:
         if d:
-            result.update(d)
+            for k, v in d.items():
+                # delete cleared key
+                if isinstance(v, ClearedValue) and v.value is None:
+                    result.pop(k, None)
+
+                # just set the value
+                else:
+                    result[k] = _strip_clear_tag(v)
 
     return result
 
 
 def combine_envs(*envs):
     """Combine zero or more dictionaries containing environment variables.
+    Environment variable values may be wrapped in :py:class:`ClearedValue`.
 
     Environment variables later from dictionaries later in the list take
-    priority over those earlier in the list. For variables ending with
-    ``PATH``, we prepend (and add a colon) rather than overwriting.
+    priority over those earlier in the list.
 
-    If you pass in ``None`` in place of a dictionary, it will be ignored.
+    For variables ending with ``PATH``, we prepend (and add a colon) rather
+    than overwriting. Wrapping a path value in :py:class:`ClearedValue`
+    disables this behavior.
+
+    Environment set to ``ClearedValue(None)`` will *delete* environment
+    variables earlier in the list, rather than setting them to ``None``.
+
+    If you pass in ``None`` in place of a dictionary in **envs**, it will be
+    ignored.
     """
     return _combine_envs_helper(envs, local=False)
 
@@ -370,11 +561,19 @@ def _combine_envs_helper(envs, local):
     result = {}
     for env in envs:
         if env:
-            for key, value in env.items():
-                if key.endswith('PATH') and result.get(key):
-                    result[key] = value + pathsep + result[key]
+            for k, v in env.items():
+                # delete cleared keys
+                if isinstance(v, ClearedValue) and v.value is None:
+                    result.pop(k, None)
+
+                # append paths
+                elif (k.endswith('PATH') and result.get(k) and
+                      not isinstance(v, ClearedValue)):
+                    result[k] = v + pathsep + result[k]
+
+                # just set the value
                 else:
-                    result[key] = value
+                    result[k] = _strip_clear_tag(v)
 
     return result
 
@@ -388,7 +587,11 @@ def combine_paths(*paths):
 def combine_path_lists(*path_seqs):
     """Concatenate the given sequences into a list. Ignore None values.
     Resolve ``~`` (home dir) and environment variables, and expand globs
-    that refer to the local filesystem."""
+    that refer to the local filesystem.
+
+    .. versionchanged:: 0.4.6
+       Can take single strings as well as lists.
+    """
     results = []
 
     for path in combine_lists(*path_seqs):
@@ -410,19 +613,23 @@ def combine_opts(combiners, *opts_list):
                       combine options by that name. By default, we combine
                       options using :py:func:`combine_values`.
     :param opts_list: one or more dictionaries to combine
+
+    The dict in *opts_list* may not be wrapped in :py:class:`ClearedValue`,
+    but their values may, in which case values of that key from previous
+    opt dicts will be ignored.
     """
     final_opts = {}
 
     keys = set()
     for opts in opts_list:
-        if opts:
+        if isinstance(opts, ClearedValue):
+            raise TypeError
+        elif opts:
             keys.update(opts)
 
     for key in keys:
-        values = []
-        for opts in opts_list:
-            if opts and key in opts:
-                values.append(opts[key])
+        values = _resolve_clear_tags_in_list(
+            opts[key] for opts in opts_list if opts and key in opts)
 
         combine_func = combiners.get(key) or combine_values
         final_opts[key] = combine_func(*values)
@@ -433,7 +640,7 @@ def combine_opts(combiners, *opts_list):
 ### PRIORITY ###
 
 
-# TODO 0.4: Move inside OptionStore
+# TODO: Move inside OptionStore
 def calculate_opt_priority(opts, opt_dicts):
     """Keep track of where in the order opts were specified,
     to handle opts that affect the same thing (e.g. ec2_*instance_type).